+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
 ### Added
+- More installer validation checks for map write access and MPQEditor presence.
 
 ### Changed
 - Balance tweaks to harvesting logic. (jzy-chitong56)